+++ server/app.local.js
@@ -1,6 +1,11 @@
 var GLOBAL_CONFIG = require('../global-config');
 
+var isDevEnv = (process.env.NODE_ENV || 'development') === 'development';
+
 module.exports = {
   restApiRoot: GLOBAL_CONFIG.restApiRoot,
   livereload: process.env.LIVE_RELOAD,
+  isDevEnv: isDevEnv,
+  indexFile: require.resolve(isDevEnv ?
+    '../ngapp/index.html' : '../ngapp/dist/index.html'),
 };
