+++ server/boot/rest-api.js
@@ -1,0 +1,5 @@
+module.exports = function(app) {
+  var restApp = require('../../rest');
+  var restApiRoot = app.get('restApiRoot');
+  app.use(restApiRoot, restApp);
+};
